requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
faiss-cpu
redis
prometheus-client
joblib